Make shader compilation from non-GPU code easier
